Migrated episode availability logic into util
