--- src/app/lib/utilities/episodeStatusConst/index.js
@@ -1,7 +1,0 @@
-const episodeStatus = {
-  EPISODE_IS_AVAILABLE: 'available',
-  EPISODE_IS_NOT_YET_AVAILABLE: 'not-yet-available',
-  EPISODE_IS_EXPIRED: 'expired',
-};
-
-module.exports = episodeStatus;
